Add link to new BP en ligne
BP en ligne has a new web address, and has at my request implemented URIs that lack ampersands, to permit the modification of PN XSLT, to generate hyperlinks to BP from our biblio records. 

This commit defines a variable `$bpIdno` as the value of `//t:idno[@type='bp']` and then calls that variable twice: once in html's `a/@href` (where it is concatenated with the new web address endpoint, as its parameter), and once as the value of that `a/@href` element.